services/training/requirements.txt
pandas==2.2.2
numpy==1.26.4
statsmodels==0.14.2
lightgbm==4.3.0
mlflow==2.13.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1
matplotlib==3.9.0
